仓库调拨-数量支持小数

+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/stock/transfer/QueryScTransferOrderBo.java
@@ -64,7 +64,7 @@ public class QueryScTransferOrderBo extends BaseBo<ScTransferOrder> {
    * 调拨数量
    */
   @ApiModelProperty("调拨数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 调拨成本金额
