+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/stock/transfer/QueryScTransferOrderDetailReceiveBo.java
@@ -3,6 +3,7 @@ package com.lframework.xingyun.sc.bo.stock.transfer;
 import com.lframework.starter.web.core.bo.BaseBo;
 import com.lframework.xingyun.sc.entity.ScTransferOrderDetailReceive;
 import io.swagger.annotations.ApiModelProperty;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Data;
 
@@ -19,7 +20,7 @@ public class QueryScTransferOrderDetailReceiveBo extends BaseBo<ScTransferOrderD
    * 收货数量
    */
   @ApiModelProperty("收货数量")
-  private Integer receiveNum;
+  private BigDecimal receiveNum;
 
   /**
    * 创建人
