+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/ScTransferOrder.java
@@ -47,7 +47,7 @@ public class ScTransferOrder extends BaseEntity implements BaseDto {
   /**
    * 调拨数量
    */
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 调拨成本金额
