+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/impl/stock/transfer/ScTransferOrderDetailServiceImpl.java
@@ -4,6 +4,7 @@ import com.lframework.starter.web.core.impl.BaseMpServiceImpl;
 import com.lframework.xingyun.sc.entity.ScTransferOrderDetail;
 import com.lframework.xingyun.sc.mappers.ScTransferOrderDetailMapper;
 import com.lframework.xingyun.sc.service.stock.transfer.ScTransferOrderDetailService;
+import java.math.BigDecimal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,7 +15,7 @@ public class ScTransferOrderDetailServiceImpl
 
   @Transactional(rollbackFor = Exception.class)
   @Override
-  public int receive(String orderId, String productId, Integer receiveNum) {
+  public int receive(String orderId, String productId, BigDecimal receiveNum) {
     return getBaseMapper().receive(orderId, productId, receiveNum);
   }
