+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/service/stock/transfer/ScTransferOrderDetailService.java
@@ -2,6 +2,7 @@ package com.lframework.xingyun.sc.service.stock.transfer;
 
 import com.lframework.starter.web.core.service.BaseMpService;
 import com.lframework.xingyun.sc.entity.ScTransferOrderDetail;
+import java.math.BigDecimal;
 
 public interface ScTransferOrderDetailService extends BaseMpService<ScTransferOrderDetail> {
 
@@ -12,7 +13,7 @@ public interface ScTransferOrderDetailService extends BaseMpService<ScTransferOr
    * @param receiveNum
    * @return
    */
-  int receive(String orderId, String productId, Integer receiveNum);
+  int receive(String orderId, String productId, BigDecimal receiveNum);
 
   /**
    * 统计未收货的商品
